+++ src/corpus/corpus.composable.ts
@@ -100,14 +100,7 @@ export function useCorpus(corpusId: string) {
   async function downloadSource(source: FileMeta, binary: boolean) {
     return spin(
       api.downloadSources(corpusId, source.name, binary),
-      `corpus/${corpusId}/sources/${source.name}/raw`,
-    );
-  }
-
-  async function downloadPlaintext(source: FileMeta) {
-    return spin(
-      api.downloadSourceText(corpusId, source.name),
-      `corpus/${corpusId}/sources/${source.name}/plain`,
+      `corpus/${corpusId}/sources/${source.name}`,
     );
   }
 
@@ -173,7 +166,6 @@ export function useCorpus(corpusId: string) {
     sources,
     hasSources,
     downloadSource,
-    downloadPlaintext,
     uploadSources,
     deleteSource,
     extensions,